Backend/requirements.txt
openai
python-dotenv
Flask
requests